requirements.txt
django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
pillow>=8.2.0,<8.3.0
uwsgi>=2.0.19,<2.1
djangorestframework
drf-spectacular
Pillow
whitenoise